pyproject.toml
[project]
name = "mlops-assesment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "uvicorn>=0.34.3",
]
